Update Project.ipynb
I did the pretty basic part of the project in this code described above. The lines carry out the following activities:

- Set all Directories (I could not do it in a simpler way, I believe we can do something better)
- Extract ZipFiles:
       a.Images Zipfiles and;
       b.Metadata Zipfile
- Load all the images in a list.
- Turn the list into a numpy array.
- Load the metadata
- Verify that the image and metadata objects are in the same order and deal with the same image / metadata.
- And to finish I did one-hot encoding of the metadata.

Now I will work on improving the loading of images. So far the loading is with the raw images, without any treatment. I studied a little of the subject and I think I can do something much better. And we will also have to reduce the size of the images, because when loading all 10015 images are consumed about 9 to 10 Gb of memory. It is soon impossible to work with this file size.